+++ plugins/harlequin_duckdb/adapter.py
@@ -1,16 +1,15 @@
 from __future__ import annotations
 
 import json
-from abc import ABC, abstractmethod
 from pathlib import Path
-from typing import Any, Final, Sequence
+from typing import Sequence
 
 import duckdb
 from duckdb.typing import DuckDBPyType
-from textual_fastdatatable.backend import AutoBackendType
-
+from harlequin.adapter import HarlequinAdapter, HarlequinConnection, HarlequinCursor
 from harlequin.catalog import Catalog, CatalogItem
 from harlequin.exception import (
+    HarlequinConnectionError,
     HarlequinCopyError,
     HarlequinQueryError,
 )
@@ -20,161 +19,12 @@ from harlequin.export_options import (
     JSONOptions,
     ParquetOptions,
 )
-from harlequin.options import HarlequinAdapterOption, HarlequinCopyOptions
+from textual_fastdatatable.backend import AutoBackendType
 
 
-class HarlequinCursor(ABC):
-    @abstractmethod
-    def __init__(self, *args: Any, **kwargs: Any) -> None:
-        pass
-
-    @abstractmethod
-    def columns(self) -> list[tuple[str, str]]:
-        """
-        Gets a list of columns for the result set of the cursor.
-
-        Returns: list[tuple[str, str]], where each tuple is the (name, type) of
-            a column, where type should be a short (1-3 character) string. The
-            columns must be ordered in the same order as the data returned by
-            fetchall()
-        """
-        pass
-
-    @abstractmethod
-    def set_limit(self, limit: int) -> "HarlequinCursor":
-        """
-        Limits the number of results for future calls to fetchall().
-
-        Args:
-            limit (int): The maximum number of records to be returned
-            by future calls to fetchall().
-
-        Returns: HarlequinCursor, either a reference to self or a new
-            cursor with the limit applied.
-        """
-        pass
-
-    @abstractmethod
-    def fetchall(self) -> AutoBackendType:
-        """
-        Returns data from the cursor's result set. Can return any type supported
-        by textual-fastdatatable.
-
-        Returns:
-            pyarrow.Table |
-            pyarrow.Record Batch |
-            Sequence[Iterable[Any]] |
-            Mapping[str, Sequence[Any]]
-        """
-        pass
-
-
-class HarlequinConnection(ABC):
-    @abstractmethod
-    def __init__(self, *args: Any, **kwargs: Any) -> None:
-        pass
-
-    @abstractmethod
-    def execute(self, query: str) -> HarlequinCursor | None:
-        """
-        Executes query and returns a cursor (for a select stmt) or None. Raises
-        HarlequinQueryError if the database raises an error in response to the query.
-
-        Args:
-            query (str): The text of a single query to execute
-
-        Returns: HarlequinCursor | None
-
-        Raises: HarlequinQueryError
-        """
-        pass
-
-    @abstractmethod
-    def get_catalog(self) -> Catalog:
-        """
-        Introspects the connected database and returns a Catalog object with
-        items for each database, schema, table, view, column, etc.
-
-        Returns: Catalog
-        """
-        pass
-
-    def copy(self, query: str, path: Path, options: ExportOptions) -> None:
-        """
-        Exports data returned by query to a file or directory at path, using
-        options.
-        Args:
-            query (str): The text of the query (select stmt) to be executed.
-            path (Path): The destination location for the file(s) to be written.
-            options (HarlequinCopyOptions): An instance of copy options specific to
-                this operation and adapter.
-
-        Returns: None
-
-        Raises:
-            NotImplementedError if the adapter does not have copy functionality.
-            HarlequinCopyError for all other exceptions during export.
-        """
-        raise NotImplementedError
-
-    def validate_sql(self, text: str) -> str:
-        """
-        Parses text as one or more queries; returns text if parsing does not result
-        in an error; otherwise returns the empty string ("").
-
-        Args:
-            text (str): The text, which may compose one or more queries and partial
-                queries.
-
-        Returns: str, either the original text or the empty string ("")
-
-        Raises: NotImplementedError if the adapter does not provide this optional
-            functionality.
-        """
-        raise NotImplementedError
-
-
-class HarlequinAdapter(ABC):
-    """
-    A HarlequinAdapter is the main abstraction for a database backend for
-    Harlequin. It must declare its configuration setting the ADAPTER_OPTIONS
-    class variable. If the adapter supports copying (exporting
-    data to a file or directory), it also must declare COPY_OPTIONS.
-    """
-
-    ADAPTER_OPTIONS: Final[list[HarlequinAdapterOption] | None] = None
-    COPY_OPTIONS: Final[HarlequinCopyOptions | None] = None
-
-    @abstractmethod
-    def __init__(self, conn_str: Sequence[str], **options: Any) -> None:
-        pass
-
-    @abstractmethod
-    def connect(self) -> tuple[HarlequinConnection, str]:
-        """
-        Creates and returns an initialized connection to a database. Necessary config
-        should be stored in the HarlequinAdapter instance when it is created.
-
-        Returns: tuple[HarlequinConnection, str], where the str is a message that
-            will be passed to the user in a notification. If str is the empty string,
-            no notification will be presented to the user.
-
-        Raises: HarlequinConnectionError if a connection could not be established.
-        """
-        pass
-
-    @property
-    def implements_copy(self) -> bool:
-        """
-        True if the adapter's connection implements the copy() method. Adapter must
-        also provide options for customizing the Export dialog GUI.
-        """
-        return self.COPY_OPTIONS is not None
-
-
-class HarlequinDuckDBCursor(HarlequinCursor):
+class DuckDbCursor(HarlequinCursor):
     def __init__(
-        self, conn: HarlequinDuckDBConnection, relation: duckdb.DuckDBPyRelation
+        self, conn: DuckDbConnection, relation: duckdb.DuckDBPyRelation
     ) -> None:
         self.conn = conn
         self.relation = relation
@@ -204,7 +54,7 @@ class HarlequinDuckDBCursor(HarlequinCursor):
         return result  # type: ignore
 
 
-class HarlequinDuckDBConnection(HarlequinConnection):
+class DuckDbConnection(HarlequinConnection):
     RELATION_TYPE_MAPPING = {
         "BASE TABLE": "t",
         "LOCAL TEMPORARY": "tmp",
@@ -332,7 +182,7 @@ class HarlequinDuckDBConnection(HarlequinConnection):
                     ),
                 ) from e
 
-    def execute(self, query: str) -> HarlequinDuckDBCursor | None:
+    def execute(self, query: str) -> DuckDbCursor | None:
         try:
             rel = self.conn.sql(query)
         except duckdb.Error as e:
@@ -342,7 +192,7 @@ class HarlequinDuckDBConnection(HarlequinConnection):
             ) from e
 
         if rel is not None:
-            return HarlequinDuckDBCursor(conn=self, relation=rel)
+            return DuckDbCursor(conn=self, relation=rel)
         else:
             return None
 
@@ -471,3 +321,152 @@ class HarlequinDuckDBConnection(HarlequinConnection):
         return cls.COLUMN_TYPE_MAPPING.get(
             str(native_type).split("(")[0], cls.UNKNOWN_TYPE
         )
+
+
+class DuckDbAdapter(HarlequinAdapter):
+    def __init__(
+        self,
+        conn_str: Sequence[str],
+        init_path: Path | None = None,
+        no_init: bool = False,
+        read_only: bool = False,
+        allow_unsigned_extensions: bool = False,
+        extension: list[str] | None = None,
+        force_install_extensions: bool = False,
+        custom_extension_repo: str | None = None,
+        md_token: str | None = None,
+        md_saas: bool = False,
+    ) -> None:
+        self.conn_str = conn_str if conn_str else (":memory:",)
+        self.init_path = init_path
+        self.no_init = no_init
+        self.read_only = read_only
+        self.allow_unsigned_extensions = allow_unsigned_extensions
+        self.extensions = extension if extension is not None else []
+        self.force_install_extensions = force_install_extensions
+        self.custom_extension_repo = custom_extension_repo
+        self.md_token = md_token
+        self.md_saas = md_saas
+
+    def connect(self) -> tuple[DuckDbConnection, str]:
+        primary_db, *other_dbs = self.conn_str
+        token = f"?token={self.md_token}" if self.md_token else ""
+        saas = "?saas_mode=true" if self.md_saas else ""
+        config = {
+            "allow_unsigned_extensions": str(self.allow_unsigned_extensions).lower()
+        }
+
+        try:
+            connection = duckdb.connect(
+                database=f"{primary_db}{token}{saas}",
+                read_only=self.read_only,
+                config=config,
+            )
+            for db in other_dbs:
+                connection.execute(
+                    f"attach '{db}'{' (READ_ONLY)' if self.read_only else ''}"
+                )
+        except (duckdb.CatalogException, duckdb.IOException) as e:
+            raise HarlequinConnectionError(
+                str(e), title="DuckDB couldn't connect to your database."
+            ) from e
+
+        if self.custom_extension_repo:
+            connection.execute(
+                f"SET custom_extension_repository='{self.custom_extension_repo}';"
+            )
+
+        if self.extensions:
+            try:
+                for extension in self.extensions:
+                    # todo: support installing from a URL instead.
+                    connection.install_extension(
+                        extension=extension, force_install=self.force_install_extensions
+                    )
+                    connection.load_extension(extension=extension)
+            except (duckdb.HTTPException, duckdb.IOException) as e:
+                raise HarlequinConnectionError(
+                    str(e), title="DuckDB couldn't install or load your extension."
+                ) from e
+
+        msg = ""
+        if self.init_path is not None and not self.no_init:
+            init_script = self._read_init_script(self.init_path)
+            try:
+                count = 0
+                for command in self._split_script(init_script):
+                    rewritten_command = self._rewrite_init_command(command)
+                    for cmd in rewritten_command.split(";"):
+                        if cmd.strip():
+                            connection.execute(cmd)
+                            count += 1
+            except duckdb.Error as e:
+                msg = f"Attempted to execute script at {init_script[0]}\n{e}"
+                raise HarlequinConnectionError(
+                    msg, title="DuckDB could not execute your initialization script."
+                ) from e
+            else:
+                if count > 0:
+                    msg = f"Executed {count} commands from {self.init_path}"
+
+        return DuckDbConnection(conn=connection), msg
+
+    @staticmethod
+    def _read_init_script(init_path: Path) -> str:
+        try:
+            with open(init_path.expanduser(), "r") as f:
+                init_script = f.read()
+        except OSError:
+            init_script = ""
+        return init_script
+
+    @staticmethod
+    def _split_script(script: str) -> list[str]:
+        """
+        DuckDB init scripts can contain SQL queries or dot commands. The SQL
+        queries may contain newlines, but the dot commands are newline-terminated.
+        This takes a raw script and returns a list of executable commands
+        """
+        lines = script.splitlines()
+        commands: list[str] = []
+        i = 0
+        for j, line in enumerate(lines):
+            if line.startswith("."):
+                commands.append("\n".join(lines[i:j]))
+                commands.append(line)
+                i = j + 1
+        commands.append("\n".join(lines[i:]))
+        return [command.strip() for command in commands if command]
+
+    @staticmethod
+    def _rewrite_dot_open(command: str) -> str:
+        """
+        Rewrites .open command into its SQL equivalent.
+        """
+        args = command.split()[1:]
+        if not args:
+            return "attach ':memory:'; use memory;"
+        else:
+            # --readonly is only supported option
+            if len(args) == 2 and args[0] == "--readonly":
+                option = " (READ_ONLY)"
+                db_path = Path(args[1])
+            else:
+                option = ""
+                db_path = Path(args[0])
+
+            return f"attach '{db_path}'{option} as {db_path.stem}; use {db_path.stem};"
+
+    @classmethod
+    def _rewrite_init_command(cls, command: str) -> str:
+        """
+        DuckDB init scripts can contain dot commands, which can only be executed
+        by the CLI, not the python API. Here, we rewrite some common ones into
+        SQL, and rewrite the others to be no-ops.
+        """
+        if not command.startswith("."):
+            return command
+        elif command.startswith(".open"):
+            return cls._rewrite_dot_open(command)
+        else:
+            return ""