+++ tests/conftest.py
@@ -1,9 +1,17 @@
+from __future__ import annotations
+
+import sys
 from pathlib import Path
 
 import duckdb
 import pytest
 from harlequin import Harlequin
-from harlequin.adapter import DuckDBAdapter
+from harlequin.adapter import HarlequinAdapter
+
+if sys.version_info < (3, 10):
+    from importlib_metadata import entry_points
+else:
+    from importlib.metadata import entry_points
 
 
 @pytest.fixture
@@ -39,15 +47,26 @@ def small_duck(tmp_path: Path, data_dir: Path) -> Path:
 
 
 @pytest.fixture
-def app() -> Harlequin:
-    return Harlequin(DuckDBAdapter([":memory:"], no_init=True))
+def duckdb_adapter() -> type[HarlequinAdapter]:
+    eps = entry_points(group="harlequin.adapter")
+    cls: type[HarlequinAdapter] = eps["duckdb"].load()  # type: ignore
+    return cls
 
 
 @pytest.fixture
-def app_small_duck(small_duck: Path) -> Harlequin:
-    return Harlequin(DuckDBAdapter([str(small_duck)], no_init=True))
+def app(duckdb_adapter: type[HarlequinAdapter]) -> Harlequin:
+    return Harlequin(duckdb_adapter([":memory:"], no_init=True))
 
 
 @pytest.fixture
-def app_multi_duck(tiny_duck: Path, small_duck: Path) -> Harlequin:
-    return Harlequin(DuckDBAdapter([str(tiny_duck), str(small_duck)], no_init=True))
+def app_small_duck(
+    duckdb_adapter: type[HarlequinAdapter], small_duck: Path
+) -> Harlequin:
+    return Harlequin(duckdb_adapter([str(small_duck)], no_init=True))
+
+
+@pytest.fixture
+def app_multi_duck(
+    duckdb_adapter: type[HarlequinAdapter], tiny_duck: Path, small_duck: Path
+) -> Harlequin:
+    return Harlequin(duckdb_adapter([str(tiny_duck), str(small_duck)], no_init=True))